# perf(pnp): lazy initialize libzip