**What's the problem this PR addresses?**

The PnP runtime always initializes Libzip even when it isn't going to be used, for example when `@yarnpkg/core` is asking `@yarnpkg/plugin-pnp` where packages are located.

Ref https://github.com/yarnpkg/berry/issues/2575

**How did you fix it?**

Lazy initialize libzip

**Result**

Tested on the Storybook repo using PnP

```
$ YARN_IGNORE_PATH=1 nodejs before.js && YARN_IGNORE_PATH=1 hyperfine -w 5 "nodejs before.js echo"
Benchmark #1: nodejs before.js echo
  Time (mean ± σ):      1.837 s ±  0.043 s    [User: 1.451 s, System: 0.209 s]
  Range (min … max):    1.804 s …  1.955 s    10 runs

$ YARN_IGNORE_PATH=1 nodejs after.js && YARN_IGNORE_PATH=1 hyperfine -w 5 "nodejs after.js echo"
Benchmark #1: nodejs after.js echo
  Time (mean ± σ):      1.750 s ±  0.010 s    [User: 1.304 s, System: 0.224 s]
  Range (min … max):    1.741 s …  1.775 s    10 runs
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.