@arcanis 2021-05-24
I tried a little change to avoid storing different value types (object / functions) inside the same class property, as that sometimes can have adverse effects on engine optimizations (and this way each property has a single concern).